+++ config/environments/development.rb
@@ -35,6 +35,11 @@ Rails.application.configure do
 
   # config.action_mailer.perform_caching = false
 
+  # For graphiql
+  config.session_store :cookie_store, key: '_interslice_session'
+  config.middleware.use ActionDispatch::Cookies
+  config.middleware.use config.session_store, config.session_options
+
   # Print deprecation notices to the Rails logger.
   config.active_support.deprecation = :log