+++ CHANGELOG.md
@@ -6,6 +6,7 @@
     - `rails db:migrate`
   - [5-user-can-manage-courses-with-lessons](https://trello.com/c/W2Z7m1At)
     - `rails db:migrate`
+  - [6-build-graphql-courses-schema](https://trello.com/c/FnUmdkuD)
 - Improvements
   - [7-respond-sections-and-lessons-datum-with-eager-loading-on-course-index-api-endpoints](https://trello.com/c/J2ivwOIJ)
 - Bug fixes